requirements.txt
pandas==1.4.2
prometheus-client==0.8.0
py4j==0.10.9
pyodbc==4.0.30
pyspark==3.1.2
python-dateutil==2.8.1
PyYAML==6.0
requests==2.24.0
SQLAlchemy==1.4.36
chispa==0.8.2
pytest==6.2.5
pylint==2.12.2
behave==1.2.6
pyarrow>=1.0.0
pytest-cov==3.0.0